[refactor]: 네이밍과 요청 url 변경
